выносит функцию для рендера в отдельный файл
